refactor: rename User to Member
