+++ src/main/java/com/likelion/oegaein/domain/member/Gender.java
@@ -1,0 +1,5 @@
+package com.likelion.oegaein.domain.member;
+
+public enum Gender {
+    M, F
+}
