+++ src/main/java/com/likelion/oegaein/domain/member/Mbti.java
@@ -1,4 +1,4 @@
-package com.likelion.oegaein.domain.user;
+package com.likelion.oegaein.domain.member;
 
 public enum Mbti {
     ESFJ,
