--- src/main/java/com/likelion/oegaein/domain/user/Gender.java
@@ -1,5 +1,0 @@
-package com.likelion.oegaein.domain.user;
-
-public enum Gender {
-    M, F
-}
