+++ src/main/java/com/likelion/oegaein/repository/MemberRepository.java
@@ -1,0 +1,22 @@
+package com.likelion.oegaein.repository;
+
+import com.likelion.oegaein.domain.member.Member;
+import jakarta.persistence.EntityManager;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+
+@Repository
+@RequiredArgsConstructor
+public class MemberRepository {
+
+    private final EntityManager em;
+
+    public Long save(Member user){
+        em.persist(user);
+        return user.getId();
+    }
+
+    public Member find(Long id){
+        return em.find(Member.class, id);
+    }
+}
